Started working on the parser.
Pattern parsing does not work. Currently, it just parses anything
within the [A-Za-z] range as a variable. It can parse rules of the
following form:

    A | B | C = D = E

That is, zero or more guards followed by two or more equal
patterns. It will also parse multiple lines, with any amount of
whitespace in between.

Even with the limitations, it's a great start, and it's all downhill
from here.